refactor: Separate MediaUploader constants export
This allows the MediaUploader component file to export only React
components, which enables hot module reloading.
